redesigned delete cart

+++ src/Shopping/delete_cart.php
@@ -1,26 +1,34 @@
 <?php  
 
-    session_start();
+    function delete_cart_user(){
 
-    $path = $_SERVER['DOCUMENT_ROOT'];
-    $path .= "/database.php";
-    include_once($path);
+      session_start();
 
-    $cart_id = $_SESSION["cart_id"];
+      //Create connection to database
+      $path = $_SERVER['DOCUMENT_ROOT'];
+      $path .= "/database.php";
+      include_once($path);
 
-    $query = $con->prepare("DELETE FROM CART_ITEMS WHERE cart_id=?");
-    $query->bind_param("i", $cart_id);
-    $query->execute();
-    $query->close();
+      $cart_id = $_SESSION["cart_id"];
 
-    $query = $con->prepare("DELETE FROM CARTS WHERE cart_id=?");
-    $query->bind_param("i", $cart_id);
-    $query->execute();
-    $query->close();
+      //If the user have a cart delete it. 
+      if(gettype($_SESSION["cart_id"]) != "NULL" && isset($_SESSION["cart_id"]) )
+    
+        $query = $con->prepare("DELETE FROM CART_ITEMS WHERE cart_id=?");
+        $query->bind_param("i", $cart_id);
+        $query->execute();
+        $query->close();
+    
+        $query = $con->prepare("DELETE FROM CARTS WHERE cart_id=?");
+        $query->bind_param("i", $cart_id);
+        $query->execute();
+        $query->close();
+    
+        unset($_SESSION["cart_id"]);//Reset cart_id variable
 
-    unset($_SESSION["cart_id"]);//Reset cart_id variable
+        header("Location: /Shopping/shopping_cart.php");
+        exit;
 
-    header("Location: shopping_cart.php");
-
+    }
 
 ?>
